Using MiniEngine's color buffer for raytracing resource
This took WAY longer than expected. Most examples were using dynamic binding for CPU HANDLES which is was not available for Descriptor Tables types. Additionally, the GetUAV function was returning a const GPU HANDLE. In the end the GPU  HANDLE was acquired from the Descriptor Heap's [] operator. The issues was made worse by the fact that mistakingly theSRV resource handle was supplied