Fixed surfel list bug leading to disappearing surfels
The surfel list was not big enough to hold the ranges of the surfels inside. Prefix sum was working correcly but was giving indices of ranged that were not mapped leading to some cell data looking like its slowly disappearing. The issue was fixed by adjusting the surfel list count to be _CELL_COUNT_ x _SURFEL_PER_CELL